Improve version strings a bit
Instead of ".dev" followed by number of commits from the recent version
tag, use the number of commits from it that are on master, then "+"
followed by the number of commits since the branching point and the
usual ".g<sha1>", finnaly followed by an optional ".local" and ".dirty".
This makes it easier to identify older build materials that happen to
have lots of commits from new ones that have few commits.  The main
difference is in the "+" part so it shouldn't have any bad effects.

Also, be a little more strict in verifying tag versions: forbid a third
"0" component.