Move Python Class docs to text files
The class-level descriptions for the generated python classes that wrap
the scala classes were contained in the `WrapperClassDoc.scala` file as
formatted strings that were selected in a `case` statement based on
`ClassName`.  There were two major disadvantages to this: (1) the file
has to be edited and Codegen has to be rebuilt every time a new class is
added and (2) developers are likely to skip this step because the
documentation does not live with the code.  The text of the formatted
strings has been moved to text files that are part of the projects that
they live with.  Now `WrapperClassDoc.scala` reads the text file to
provide the wrapper generator with the class doc string for the python.

Reformatted the docstring for `ImageReader.py` to keep most line to 80
characters of less for readability.