Patch the Spark jars to make WASB access work
Ideally, we would just add `hadoop-azure` to the SBT dependencies, but
that brings in a some hadoop classes that are incompatible with the ones
that are included in Spark (since it uses an older hadoop version).
Instead, get the jar file directly into `lib/spark/jars`, and do the
same for `azure-storage` which is a dependency.