Sample notebook additions
Sample notebook 104 (based on Sample 6 from the Studio AzureML Gallery),
and 105 (which is an alternate version of 102).  These notebooks
demonstrate MMLSpark APIs which haven't appeared in other sample
notebooks.

Notebook 104 uses `Pipeline` stages for model fitting.  The stages are the
CleanMissingData and the regressor.

Notebook 105 uses `DataConversion` to convert categorical columns,
eliminating the need to iterate over the columns to be converted.