Use a "0.0" version for local builds
This, in addition to the previous commit, greatly simplifies dealing
with versions in local builds and make local building less fragile.  The
only downside of not being able to easily distribute build artifacts of
such builds (but nobody ever did that in practice).

Also, add nullglob to the bash settings; makes things more sane, and
shouldn't break anything... except the surprising result when `unset` is
used with unquoted array references which is fixed (see
https://stackoverflow.com/questions/15897473).