Remove libraries that are no longer needed
These libraries were mainly needed for CNTK which didn't have them, but
it was fixed in v2.4, so they are no longer needed.  (Still needs to be
done for the GPU version, when it's bumped to 2.4.)

Leave `libopenmpi1.10`, `libgomp1`, and `libunwind8` in the plus image,
in case it's needed there.