Update ComputeModelStatistics.scala by converting label and prediction to double type
When the `label` or `prediction` are not both of double type, it throws the "error prediction and label columns invalid or missing". 

To users, this error is not informative as one can spend hours trying to figure out the cause of the error despite the 'label' and 'prediction' variables being present in the data frame.

I propose converting the label and prediction columns to double type when they are both present in the dataframe.