chore: Adding responseFormat parameter in OpenAI Chat Completion Request (#2329)
* chore: Adding responseFormat parameter in OpenAI Chat Completion Request

OpenAI chat completion accepts `response_format` as request parameter. Setting this parameter as `json_object`guarantees that response from OpenAi service will be in JSON format. It does also require that one of the prompt must also have the word JSON in it. In this PR, we are adding a new field in request, which can be `json_object` or `text` as prescribed by OpenAI. In case it is set as `json_object`, then we also add a system prompt instructing openAI to respond in JSON format.

Unit tests have been added to validate this functionality

* removing some extra lines to conform to style guidelines

* reducing the max token count to enable tests to pass

* moving the functionality to OpenAIPrompt class

* Fixing failing unit tests

---------

Co-authored-by: Farrukh Masud <farrukhmasud@microsoft.com>