chore: Remove redundant text message for file input in OpenAIPrompt
When using file input in OpenAIPrompt before, a text message will be sent along: The name of the file to analyze is %s.\nHere is the content:\n, this will cause info leak in file name, as long as extra token used. So it's for the best to remove this piece of message.

authored-by: Wendong Li <wendongli@microsoft.com>